refactor: rename interns1_pro_ts.py to interns1_pro_time_series.py
